20-50kHz sweep working

+++ firmware/src/pwm/pwm.h
@@ -15,7 +15,7 @@
 #define pwm_buf_size 2048
 
 void print_lut(uint16_t pwm_num_samples);
-uint32_t sine_pwm_lut[pwm_buf_size];
+uint8_t sine_pwm_lut[pwm_buf_size];
 
 bool generate_sine_pwm(double amplitude, uint16_t pwm_num_samples);
 bool play_frequency(double amplitude, double frequency);
